Move OSDeviceInfoError and handler
* Move OSDeviceInfoError and OSDeviceInfoCompletionHandler to OneSignalStateSynchronizer
* Add appropriate imports in files OneSignal.java, OneSignalStateSynchronizer and UserStateSynchronizer
